bh——add chart endpoint, data collection, data table, data source

+++ project-bi-engine-core/src/main/java/com/patsnap/insights/trickydata/dao/UserDao.java
@@ -4,8 +4,6 @@ import com.patsnap.insights.trickydata.entity.UserEntity;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserDao extends CrudRepository<UserEntity, Long> {
-
-
+public interface UserDao extends CrudRepository<UserEntity, Integer> {
     public UserEntity getUserByName(String name);
 }
